pyproject.toml
[tool.poetry]
name = "planning-through-contact"
version = "0.1.0"
description = ""
authors = ["Bernhard Paus Græsdal <bernhardpg@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "~3.12"
matplotlib = "^3.6.1"
numpy = "^1.23.3"
pdbpp = "^0.10.3"
tqdm = "^4.64.1"
pydot = "^1.4.2"
tk = "^0.1.0"
drake = "^1.24"
pytest = "^8.2.0"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
pygame = "^2.5.1"
opencv-python = "^4.9.0.80"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
underactuated = "^2024.4.27"
setuptools = "^70.0.0"
black = "^24.4.2"
flake8 = "^7.0.0"
autoflake = "^2.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"
natsort = "^8.4.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# PyRight wrongly reports that several Drake modules are missing (could also be an issue with Drake)
reportMissingImports = false

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 88
skip_gitignore = true
